requirements.txt
matplotlib==3.9.2
numpy==2.1.2
opencv-python-headless==4.10.0.84
opencv-python==4.10.0.84
scikit-image==0.24.0
pytorch-metric-learning
faiss-gpu
faiss-cpu
torchvision
torchviz
neptune
torchsummary
black==24.10.0
flake8==7.1.1
pylint==3.3.1
pytest==8.3.3
pre-commit==4.0.1
